refactor: :building_construction: change the folder architecture format for the Components folder
This change will change the index.ts to [Component].tsx + index.ts as entrypoint
