+++ src/components/CustomNode/TextNode.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { MdLink, MdLinkOff } from "react-icons/md";
 // import { useInViewport } from "react-in-viewport";
-import { CustomNodeProps } from "src/components/CustomNode";
+import { CustomNodeProps } from "src/components/CustomNode/CustomNode";
 import useGraph from "src/store/useGraph";
 import useStored from "src/store/useStored";
 import styled from "styled-components";
