+++ src/components/Producthunt/index.ts
@@ -1,0 +1,1 @@
+export * from './Producthunt';
