+++ src/components/Sponsors/index.ts
@@ -1,0 +1,1 @@
+export * from './Sponsors';
