+++ src/components/Toggle/index.ts
@@ -1,0 +1,1 @@
+export { default as Toggle } from './Toggle';
