+++ src/containers/Modals/SettingsModal/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Modal, ModalProps } from "src/components/Modal";
-import Toggle from "src/components/Toggle";
+import {Toggle} from "src/components/Toggle";
 import useStored from "src/store/useStored";
 import styled from "styled-components";
 import shallow from "zustand/shallow";
